bot/requirements.txt
discord.py==2.6.3
SQLAlchemy==2.0.41
